Moving all usage of explicit flaws out of DamageModel base class.
