+++ src/Damage/DamageModelInline.hh
@@ -22,40 +22,6 @@ requireIntersectionConnectivity() const {
 }
 
 //------------------------------------------------------------------------------
-// The effective critical number of nodes per smoothing scale, below which we
-// assume all flaws are active on a node.
-//------------------------------------------------------------------------------
-template<typename Dimension>
-inline
-double
-DamageModel<Dimension>::
-criticalNodesPerSmoothingScale() const {
-  return mCriticalNodesPerSmoothingScale;
-}
-
-template<typename Dimension>
-inline
-void
-DamageModel<Dimension>::
-criticalNodesPerSmoothingScale(double x) {
-  VERIFY(x >= 0.0);
-  mCriticalNodesPerSmoothingScale = x;
-}
-
-//------------------------------------------------------------------------------
-// Return the set of flaw activation energies for the given node.
-//------------------------------------------------------------------------------
-template<typename Dimension>
-inline
-const std::vector<double>
-DamageModel<Dimension>::
-flawsForNode(const size_t index) const {
-  REQUIRE(index < mNodeList.numInternalNodes());
-  REQUIRE(mFlaws.nodeListPtr() == &mNodeList);
-  return mFlaws(index);
-}
-
-//------------------------------------------------------------------------------
 // Access the internal parameters of the model.
 //------------------------------------------------------------------------------
 template<typename Dimension>
@@ -104,44 +70,6 @@ const NodeCoupling&
 DamageModel<Dimension>::
 nodeCoupling() const {
   return *mNodeCouplingPtr;
-}
-
-//------------------------------------------------------------------------------
-// Access the state fields.
-//------------------------------------------------------------------------------
-template<typename Dimension>
-inline
-const Field<Dimension, typename Dimension::Scalar>&
-DamageModel<Dimension>::
-youngsModulus() const {
-  return mYoungsModulus;
-}
-
-template<typename Dimension>
-inline
-const Field<Dimension, typename Dimension::Scalar>&
-DamageModel<Dimension>::
-longitudinalSoundSpeed() const {
-  return mLongitudinalSoundSpeed;
-}
-
-//------------------------------------------------------------------------------
-// Access the flaw activation strains.
-//------------------------------------------------------------------------------
-template<typename Dimension>
-inline
-const typename DamageModel<Dimension>::FlawStorageType&
-DamageModel<Dimension>::
-flaws() const {
-  return mFlaws;
-}
-
-template<typename Dimension>
-inline
-typename DamageModel<Dimension>::FlawStorageType&
-DamageModel<Dimension>::
-flaws() {
-  return mFlaws;
 }
 
 }
