+++ src/Damage/TensorDamageModel.cc
@@ -15,6 +15,8 @@
 #include "TensorDamageModel.hh"
 #include "TensorStrainPolicy.hh"
 #include "TensorDamagePolicy.hh"
+#include "YoungsModulusPolicy.hh"
+#include "LongitudinalSoundSpeedPolicy.hh"
 #include "DamageGradientPolicy.hh"
 #include "Strength/SolidFieldNames.hh"
 #include "NodeList/SolidNodeList.hh"
@@ -56,7 +58,10 @@ TensorDamageModel(SolidNodeList<Dimension>& nodeList,
                   const double criticalDamageThreshold,
                   const bool damageInCompression,
                   const FlawStorageType& flaws):
-  DamageModel<Dimension>(nodeList, W, crackGrowthMultiplier, damageCouplingAlgorithm, flaws),
+  DamageModel<Dimension>(nodeList, W, crackGrowthMultiplier, damageCouplingAlgorithm),
+  mFlaws(SolidFieldNames::flaws, flaws),
+  mYoungsModulus(SolidFieldNames::YoungsModulus, nodeList),
+  mLongitudinalSoundSpeed(SolidFieldNames::longitudinalSoundSpeed, nodeList),
   mStrain(SolidFieldNames::strainTensor, nodeList),
   mEffectiveStrain(SolidFieldNames::effectiveStrainTensor, nodeList),
   mDdamageDt(TensorDamagePolicy<Dimension>::prefix() + SolidFieldNames::scalarDamage, nodeList),
@@ -138,6 +143,18 @@ registerState(DataBase<Dimension>& dataBase,
   typedef typename State<Dimension>::KeyType Key;
   typedef typename State<Dimension>::PolicyPointer PolicyPointer;
 
+  // Register Youngs modulus and the longitudinal sound speed.
+  PolicyPointer EPolicy(new YoungsModulusPolicy<Dimension>());
+  PolicyPointer clPolicy(new LongitudinalSoundSpeedPolicy<Dimension>());
+  state.enroll(mYoungsModulus, EPolicy);
+  state.enroll(mLongitudinalSoundSpeed, clPolicy);
+
+  // Set the initial values for the Youngs modulus, sound speed, and pressure.
+  typename StateDerivatives<Dimension>::PackageList dummyPackages;
+  StateDerivatives<Dimension> derivs(dataBase, dummyPackages);
+  EPolicy->update(state.key(mYoungsModulus), state, derivs, 1.0, 0.0, 0.0);
+  clPolicy->update(state.key(mLongitudinalSoundSpeed), state, derivs, 1.0, 0.0, 0.0);
+
   // Register the strain and effective strain.
   PolicyPointer effectiveStrainPolicy(new TensorStrainPolicy<Dimension>(mStrainAlgorithm));
   state.enroll(mStrain);
@@ -156,9 +173,6 @@ registerState(DataBase<Dimension>& dataBase,
   for (auto i = 0u; i < this->nodeList().numInternalNodes(); ++i) {
     if (damage(i).Trace() > mCriticalDamageThreshold) mask(i) = 0;
   }
-
-  // Register the base classes stuff.
-  DamageModel<Dimension>::registerState(dataBase, state);
 }
 
 //------------------------------------------------------------------------------
@@ -220,253 +234,60 @@ enforceBoundaries(State<Dimension>& state,
   }
 }
 
+//------------------------------------------------------------------------------
+// Cull the flaws on each node to the single weakest one.
+//------------------------------------------------------------------------------
+template<typename Dimension>
+void
+TensorDamageModel<Dimension>::
+cullToWeakestFlaws() {
+  const auto n = mFlaws.numInternalElements();
+#pragma omp parallel for
+  for (auto i = 0u; i < n; ++i) {
+    auto& flaws = mFlaws[i];
+    if (flaws.size() > 0) {
+      const auto maxVal = *max_element(flaws.begin(), flaws.end());
+      flaws = vector<double>(maxVal);
+    }
+  }
+}
 
-// //------------------------------------------------------------------------------
-// // Initialize at beginning of a step.
-// //------------------------------------------------------------------------------
-// template<typename Dimension>
-// void
-// TensorDamageModel<Dimension>::
-// initialize(const typename Dimension::Scalar& time,
-//            const typename Dimension::Scalar& dt,
-//            const DataBase<Dimension>& dataBase,
-//            State<Dimension>& state,
-//            StateDerivatives<Dimension>& derivs) {
+//------------------------------------------------------------------------------
+// Compute a Field with the sum of the activation energies per node.
+//------------------------------------------------------------------------------
+template<typename Dimension>
+Field<Dimension, typename Dimension::Scalar>
+TensorDamageModel<Dimension>::
+sumActivationEnergiesPerNode() const {
+  auto& nodeList = this->nodeList();
+  const auto n = mFlaws.numInternalElements();
+  Field<Dimension, Scalar> result("Sum activation energies", nodeList);
+#pragma omp parallel for
+  for (auto i = 0u; i < n; ++i) {
+    const auto& flaws = mFlaws(i);
+    for (const auto fij: flaws) {
+      result(i) += fij;
+    }
+  }
+  return result;
+}
 
-//   // We need to fill in the effective damage and perhaps the damage gradient.
-//   // First, grab the pertinent state.
-//   typedef typename State<Dimension>::FieldKeyType KeyType;
-//   const SolidNodeList<Dimension>& nodeList = this->nodeList();
-//   const Field<Dimension, Scalar>& mass = state.scalarField(KeyType(nodeListPtr, HydroFieldNames::mass));
-//   const Field<Dimension, Vector>& position = state.vectorField(KeyType(nodeListPtr, HydroFieldNames::position));
-//   const Field<Dimension, Scalar>& rho = state.scalarField(KeyType(nodeListPtr, HydroFieldNames::massDensity));
-//   const Field<Dimension, Scalar>& weight = state.scalarField(KeyType(nodeListPtr, HydroFieldNames::weight));
-//   const Field<Dimension, SymTensor>& H = state.symTensorField(KeyType(nodeListPtr, HydroFieldNames::H));
-//   const Field<Dimension, Scalar>& omega = state.scalarField(KeyType(nodeListPtr, HydroFieldNames::omegaGradh));
-//   const Field<Dimension, SymTensor>& D = state.symTensorField(KeyType(nodeListPtr, SolidFieldNames::tensorDamage));
-
-//   Field<Dimension, SymTensor>& Deff = state.symTensorField(KeyType(nodeListPtr, SolidFieldNames::effectiveTensorDamage));
-//   Field<Dimension, Vector>& gradD = derivs.vectorField(KeyType(&nodeList, SolidFieldNames::gradDamage));
-
-//   // Get the kernel.
-//   const TableKernel<Dimension>& W = nodeListPtr->kernel();
-//   const double etaMax2 = W.kernelExtent() * W.kernelExtent();
-
-//   // The neighbor connectivity.
-//   const ConnectivityMap<Dimension>& connectivityMap = dataBase.connectivityMap();
-//   const vector<const NodeList<Dimension>*>& nodeLists = connectivityMap.nodeLists();
-//   const size_t nodeListi = distance(nodeLists.begin(), find(nodeLists.begin(), nodeLists.end(), nodeListPtr));
-//   CHECK(nodeListi < nodeLists.size());
-
-//   const size_t firstGhostNode = nodeListPtr->firstGhostNode();
-
-//   // Iterate over the internal nodes in this NodeList.
-//   Field<Dimension, Scalar> normaliation("normalization", *nodeListPtr, 0.0);
-//   Deff = D;
-//   for (size_t i = 0; i != nodeListPtr->numInternalNodes(); ++i) {
-
-//     // State for node i.
-//     const Scalar& mi = mass(i);
-//     const Vector& ri = position(i);
-//     const Scalar& rhoi = rho(i);
-//     const Scalar& weighti = weight(i);
-//     const SymTensor& Hi = H(i);
-//     const Scalar Hdeti = Hi.Determinant();
-//     const Scalar safeOmegai = omega(i)/(omega(i)*omega(i) + 1.0e-20);
-//     const SymTensor& Di = D(i);
-//     const Scalar Dmagi = Di.Trace();
-//     CHECK(mi > 0.0);
-//     CHECK(rhoi > 0.0);
-//     CHECK(weighti > 0.0);
-//     CHECK(safeOmegai > 0.0);
-//     CHECK(Hdeti > 0.0);
-//     CHECK(Dmagi >= 0.0 and Dmagi <= Dimension::nDim);
-
-//     SymTensor& Deffi = Deff(i);
-//     Vector& gradDi = gradD(i);
-
-//     // Get the connectivity info for this node.  We only need to proceed if
-//     // there are some nodes in this list.
-//     const vector<int>& connectivity = connectivityMap.connectivityForNode(nodeListPtr, i)[nodeListi];
-//     if (connectivity.size() > 0) {
-
-//       // Iterate over the neighbors.
-//       for (vector<int>::const_iterator jItr = connectivity.begin();
-//            jItr != connectivity.end();
-//            ++jItr) {
-//         const int j = *jItr;
-
-//         // Only proceed if this node pair has not been calculated yet.
-//         if (j > i) {
-
-//           // Get the state for node j
-//           const Scalar& mj = mass(j);
-//           const Vector& rj = position(j);
-//           const Scalar& rhoj = rho(j);
-//           const Scalar& weightj = weight(j);
-//           const SymTensor& Hj = H(j);
-//           const Scalar Hdetj = Hj.Determinant();
-//           const Scalar safeOmegaj = omega(j)/(omega(j)*omega(j) + 1.0e-20);
-//           const SymTensor& Dj = D(j);
-//           const Scalar& Dmagj = Dj.Trace();
-//           CHECK(mj > 0.0);
-//           CHECK(rhoj > 0.0);
-//           CHECK(weightj > 0.0);
-//           CHECK(safeOmegaj > 0.0);
-//           CHECK(Hdetj > 0.0);
-//           CHECK(Dmagj >= 0.0 and Dmagj <= Dimension::nDim);
-
-//           SymTensor& Deffj = Deff(j);
-//           Vector& gradDnewj = gradD(j);
-
-//           // Node displacement and weighting.
-//           const Vector rij = ri - rj;
-
-//           const Vector etai = Hi*rij;
-//           const Scalar etaMagi = etai.magnitude();
-//           const std::pair<double, double> WWi = W.kernelAndGradValue(etaMagi, Hdeti);
-//           const Scalar Wi = WWi.first;
-//           const Vector gradWi = WWi.second * (Hi*etai.unitVector());
-
-//           const Vector etaj = Hj*rij;
-//           const Scalar etaMagj = etaj.magnitude();
-//           const std::pair<double, double> WWj = W.kernelAndGradValue(etaMagj, Hdetj);
-//           const Scalar Wj = WWj.first;
-//           const Vector gradWj = WWj.second * (Hj*etaj.unitVector());
-
-
-//           const Vector gradWj = W.grad(etaMagj, Hdetj) * (Hj*etaj.unitVector());
-
-//           // Increment the effective damage.
-//           switch(mEffDamageAlgorithm) {
-//           case Max:
-//             if (Dmagj > Dmaji) Di = Dj;
-//             if (Dmagi > Dmajj) Dj = Di;
-//             break;
-
-//           case Sampled:
-//             normalizationi += Dmagj*Wi;
-//             normalizationj += Dmagi*Wj;
-//             Di += Dmagj*Wi * Dj;
-//             Dj += Dmagi*Wj * Di;
-//             break;
-//           }
-
-//           // Increment the gradient.
-//           gradDi += mj*(Dj - Di)*gradWi;
-//           gradDj += mi*(Dj - Di)*gradWj;
-
-//         }
-//       }
-//     }
-
-//     // Finish the effective damage.
-//     if (mEffDamageAlgorithm == Sampled) {
-//       CHECK(normalizationi >= 0.0);
-//       Deffi /= normalizationi + tiny;
-//     }
-
-//     // Finish the gradient.
-//     if (mUseDamageGradient) {
-//       CHECK(rhoi > 0.0);
-//       gradDi /= rhoi;
-//     } else {
-//       gradDi.Zero();
-//     }
-
-//   }
-
-
-
-
-
-
-
-
-
-
-//   // If needed, compute the gradient of the damage.
-//   if (mUseDamageGradient) {
-
-//     // State fields.
-//     typedef typename State<Dimension>::FieldKeyType KeyType;
-//     const SolidNodeList<Dimension>& nodeList = this->nodeList();
-//     const Field<Dimension, Scalar>& mass = state.scalarField(KeyType(&nodeList, HydroFieldNames::mass));
-//     const Field<Dimension, Vector>& position = state.vectorField(KeyType(&nodeList, HydroFieldNames::position));
-//     const Field<Dimension, Scalar>& rho = state.scalarField(KeyType(&nodeList, HydroFieldNames::massDensity));
-//     const Field<Dimension, SymTensor>& H = state.symTensorField(KeyType(&nodeList, HydroFieldNames::H));
-//     const Field<Dimension, Scalar>& omega = state.scalarField(KeyType(nodeListPtr, HydroFieldNames::omegaGradh));
-//     const Field<Dimension, Scalar>& D = state.scalarField(KeyType(&nodeList, SolidFieldNames::tensorDamage));
-//     Field<Dimension, Vector>& gradD = derivs.vectorField(KeyType(&nodeList, SolidFieldNames::gradDamage));
-
-//     // Get the kernel.
-//     const TableKernel<Dimension>& W = nodeList.kernel();
-//     const double etaMax2 = W.kernelExtent() * W.kernelExtent();
-
-//     // The neighbor connectivity.
-//     const ConnectivityMap<Dimension>& connectivityMap = dataBase.connectivityMap();
-//     const vector<const NodeList<Dimension>*>& nodeLists = connectivityMap.nodeLists();
-//     const int nodeListi = distance(nodeLists.begin(), find(nodeLists.begin(), nodeLists.end(), &nodeList));
-//     CHECK(nodeListi < nodeLists.size());
-
-//     // Initialize the grad field.
-//     gradD.Zero();
-
-//     // Iterate over the nodes in this NodeList.
-//     for (size_t i = 0; i != nodeList.numInternalNodes(); ++i) {
-
-//       // State for node i.
-//       const Vector& ri = position(i);
-//       const Scalar& rhoi = rho(i);
-//       const SymTensor& Hi = H(i);
-//       const Scalar& Di = D(i);
-
-//       // Get the connectivity for this node.
-//       const vector< vector<int> >& fullConnectivity = connectivityMap.connectivityForNode(&nodeList, i);
-//       CHECK(fullConnectivity.size() == nodeLists.size());
-//       const vector<int>& connectivity = fullConnectivity[nodeListi];
-
-//       // Iterate over the neighbor nodes in this NodeList.
-//       for (typename vector<int>::const_iterator jItr = connectivity.begin();
-//            jItr != connectivity.end();
-//            ++jItr) {
-//         const int j = *jItr;
-     
-//         // State for node j.
-//         const Scalar& mj = mass(j);
-//         const Vector& rj = position(j);
-//         const Scalar& Dj = D(j);
-
-//         // Node displacement and weighting.
-//         const Vector rij = ri - rj;
-//         const Vector etai = Hi*rij;
-//         const Vector gradWi = Hi*W.grad(etai.magnitude(), Hi.Determinant()) * etai.unitVector();
-
-//         // Increment the gradient.
-//         gradD(i) += mj*(Dj - Di)*gradWi;
-//       }
-
-//       // Finalize the gradient.
-//       CHECK(rhoi > 0.0);
-//       gradD(i) /= rhoi;
-
-//     }
-
-//     // Apply boundary conditions to the gradient.
-//     for (ConstBoundaryIterator boundaryItr = this->boundaryBegin(); 
-//          boundaryItr != this->boundaryEnd();
-//          ++boundaryItr) {
-//       (*boundaryItr)->applyGhostBoundary(gradD);
-//     }
-//     for (ConstBoundaryIterator boundaryItr = this->boundaryBegin(); 
-//          boundaryItr != this->boundaryEnd();
-//          ++boundaryItr) {
-//       (*boundaryItr)->finalizeGhostBoundary();
-//     }
-
-//   }
-
-// }
+//------------------------------------------------------------------------------
+// Compute a Field with the number of flaws per node.
+//------------------------------------------------------------------------------
+template<typename Dimension>
+Field<Dimension, typename Dimension::Scalar>
+TensorDamageModel<Dimension>::
+numFlawsPerNode() const {
+  auto& nodeList = this->nodeList();
+  const auto n = mFlaws.numInternalElements();
+  Field<Dimension, Scalar> result("num flaws", nodeList);
+#pragma omp parallel for
+  for (auto i = 0u; i < n; ++i) {
+    result(i) = flawsForNode(i).size();
+  }
+  return result;
+}
 
 //------------------------------------------------------------------------------
 // Dump the current state to the given file.
@@ -476,6 +297,7 @@ void
 TensorDamageModel<Dimension>::
 dumpState(FileIO& file, const string& pathName) const {
   DamageModel<Dimension>::dumpState(file, pathName);
+  file.write(mFlaws, pathName + "/flaws");
   file.write(mStrain, pathName + "/strain");
   file.write(mEffectiveStrain, pathName + "/effectiveStrain");
   file.write(mDdamageDt, pathName + "/DdamageDt");
@@ -489,6 +311,7 @@ void
 TensorDamageModel<Dimension>::
 restoreState(const FileIO& file, const string& pathName) {
   DamageModel<Dimension>::restoreState(file, pathName);
+  file.read(mFlaws, pathName + "/flaws");
   file.read(mStrain, pathName + "/strain");
   file.read(mEffectiveStrain, pathName + "/effectiveStrain");
   file.read(mDdamageDt, pathName + "/DdamageDt");
